+++ crates/modor/tests/integration/app.rs
@@ -2,14 +2,14 @@ use log::Level;
 use modor::{App, FromApp, State};
 
 #[modor::test]
-fn create_node() {
+fn create_state() {
     let mut app = App::new::<Root>(Level::Info);
     app.create::<Root>();
     assert_eq!(app.get_mut::<Counter>().value, 1);
 }
 
 #[modor::test]
-fn create_node_handle() {
+fn create_state_handle() {
     let mut app = App::new::<Root>(Level::Info);
     let handle = app.handle::<Root>();
     assert_eq!(app.get_mut::<Counter>().value, 1);
@@ -18,7 +18,7 @@ fn create_node_handle() {
 }
 
 #[modor::test]
-fn take_node() {
+fn take_state() {
     let mut app = App::new::<Root>(Level::Info);
     let result = app.take(|root: &mut Root, app| {
         assert_eq!(app.get_mut::<Counter>().value, 1);
@@ -29,7 +29,7 @@ fn take_node() {
 }
 
 #[modor::test]
-fn take_node_handle() {
+fn take_state_handle() {
     let mut app = App::new::<Root>(Level::Info);
     let result = app.handle::<Root>().take(&mut app, |root: &mut Root, app| {
         assert_eq!(app.get_mut::<Counter>().value, 1);